Commit with bug parse day 2

+++ src/Services/Catalog/CatalogAPI/Models/Product.cs
@@ -13,8 +13,6 @@ namespace CatalogAPI.Models
 
         public int Description { get; set; }
 
-        public int Price { get; set; }
-
-
+        public decimal Price { get; set; }
     }
 }
